+++ pkg/await/error.go
@@ -2,7 +2,6 @@ package await
 
 import (
 	"fmt"
-
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )