+++ pkg/metadata/overrides.go
@@ -16,6 +16,7 @@ package metadata
 
 import (
 	"strconv"
+	"time"
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
@@ -25,15 +26,22 @@ func SkipAwaitLogic(obj *unstructured.Unstructured) bool {
 	return IsAnnotationTrue(obj, AnnotationSkipAwait)
 }
 
-// TimeoutSeconds returns the int value of the `pulumi.com/timeoutSeconds` annotation, or the defaultSeconds value
+// TimeoutDuration returns the timeout duration for the resource. There are a number of things it can do here in this order
+// 1. Return the timeout as specified in the customResource options
+// 2. Return the timeout as specified in `pulumi.com/timeoutSeconds` annotation,
+// 3. Return a defaultSeconds value
 // if the annotation is unset/invalid.
-func TimeoutSeconds(obj *unstructured.Unstructured, defaultSeconds int) int {
-	if s := GetAnnotationValue(obj, AnnotationTimeoutSeconds); s != "" {
+func TimeoutDuration(resourceTimeoutSeconds float64, obj *unstructured.Unstructured, defaultSeconds int) time.Duration {
+	timeout := defaultSeconds
+
+	if resourceTimeoutSeconds != 0 {
+		timeout = int(resourceTimeoutSeconds)
+	} else if s := GetAnnotationValue(obj, AnnotationTimeoutSeconds); s != "" {
 		val, err := strconv.Atoi(s)
 		if err == nil {
-			return val
+			timeout = val
 		}
 	}
 
-	return defaultSeconds
+	return time.Duration(timeout) * time.Second
 }