+++ sdk/python/pulumi_kubernetes/extensions/v1beta1/Ingress.py
@@ -30,9 +30,9 @@ class Ingress(pulumi.CustomResource):
     
     If the Ingress has not reached a Ready state after 10 minutes, it will
     time out and mark the resource update as Failed. You can override the default timeout value
-    by setting 'pulumi.com/timeoutSeconds' as a '.metadata.annotation' on the resource.
-    This approach will be deprecated in favor of customTimeouts. See
-    https://github.com/pulumi/pulumi-kubernetes/issues/672 for details.
+    by setting the 'customTimeouts' option on the resource.
+    Note that the timeout value for Delete is not yet supported. See
+    https://github.com/pulumi/pulumi-kubernetes/issues/746 for details.
     """
 
     apiVersion: pulumi.Output[str]